chore: 프리뷰 파일 분리
