+++ compose/src/main/kotlin/com/yourssu/handy/compose/Dialog.kt
@@ -13,13 +13,8 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.yourssu.handy.compose.DialogDefaults.dialogInsidePadding
 import com.yourssu.handy.compose.DialogDefaults.dialogPadding
@@ -30,7 +25,6 @@ import com.yourssu.handy.compose.button.BoxButtonSize
 import com.yourssu.handy.compose.button.BoxButtonType
 import com.yourssu.handy.compose.foundation.HandyTypography
 import com.yourssu.handy.compose.icons.HandyIcons
-import com.yourssu.handy.compose.icons.filled.Add
 import com.yourssu.handy.compose.icons.filled.Close
 
 
@@ -190,74 +184,6 @@ fun TwoButtonDialog(
                 )
             }
         }
-    }
-}
-
-@Preview
-@Composable
-private fun OneButtonDialogPreview() {
-
-    var showDialog by remember { mutableStateOf(true) }
-
-    HandyTheme {
-        Column {
-            OneButtonDialog(
-                title = "제목이 들어갑니다",
-                description = "내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다",
-                positiveText = "버튼",
-                onPositiveClick = {},
-                onDismiss = { showDialog = false } // 다이얼로그 닫기
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            OneButtonDialog(
-                title = "제목이 들어갑니다",
-                description = "내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다",
-                positiveText = "버튼",
-                onPositiveClick = {},
-                onDismiss = { showDialog = false }, // 다이얼로그 닫기
-                content = { Icon(HandyIcons.Filled.Add) }
-            )
-        }
-
-    }
-}
-
-@Preview
-@Composable
-private fun TwoButtonDialogPreview() {
-
-    var showDialog by remember { mutableStateOf(true) }
-
-    HandyTheme {
-        Column {
-
-            TwoButtonDialog(
-                title = "제목이 들어갑니다",
-                description = "내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다",
-                positiveText = "버튼",
-                onPositiveClick = {},
-                onDismiss = { showDialog = false }, // 다이얼로그 닫기
-                negativeText = "버튼",
-                onNegativeClick = {},
-            )
-
-            Spacer(modifier = Modifier.height(10.dp))
-
-            TwoButtonDialog(
-                title = "제목이 들어갑니다",
-                description = "내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다 내용이 들어갑니다",
-                positiveText = "버튼",
-                onPositiveClick = {},
-                onDismiss = { showDialog = false }, // 다이얼로그 닫기
-                negativeText = "버튼",
-                onNegativeClick = {},
-                content = { Icon(HandyIcons.Filled.Add) }
-            )
-
-        }
-
     }
 }
